move serde methods and types
to separate file
